requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
Naked==0.1.31
pycryptodome==3.15.0
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
urllib3==1.26.12
